devops_practices/big_task_linux_docker/requirements.txt
Flask==2.0.3
psycopg2-binary==2.9.6
Werkzeug==2.0.3